Add forceMultiplier parameter type and add specific field for it in parameters
