+++ newIDE/app/src/EventsSheet/EventsTree/Instruction.js
@@ -25,6 +25,7 @@ import {
   type ConnectDropTarget,
 } from 'react-dnd';
 import DropIndicator from './DropIndicator';
+import ParameterRenderingService from '../ParameterRenderingService';
 const gd = global.gd;
 const instrFormatter = gd.InstructionSentenceFormatter.get();
 instrFormatter.loadTypesFormattingFromConfig();
@@ -130,7 +131,10 @@ class Instruction extends React.Component<Props, *> {
               onClick={domEvent =>
                 this.props.onParameterClick(domEvent, parameterIndex)}
             >
-              {formattedTexts.getString(i)}
+              {ParameterRenderingService.renderParameterString(
+                parameterType,
+                formattedTexts.getString(i)
+              )}
             </span>
           );
         })}
