+++ newIDE/app/src/EventsSheet/EventsTree/Renderers/ForEachEvent.js
@@ -10,7 +10,7 @@ import {
   disabledText,
 } from '../ClassNames';
 import InlinePopover from '../../InlinePopover';
-import ObjectField from '../../InstructionEditor/ParameterFields/ObjectField';
+import ObjectField from '../../ParameterFields/ObjectField';
 import { type EventRendererProps } from './EventRenderer.flow';
 const gd = global.gd;
