+++ newIDE/app/src/EventsSheet/ParameterFields/EnumerateVariables.spec.js
