+++ newIDE/app/src/EventsSheet/ParameterFields/ExternalEventsField.js
@@ -1,21 +1,28 @@
 // @flow
 import React, { Component } from 'react';
 import AutoComplete from 'material-ui/AutoComplete';
-import { mapFor } from '../../../Utils/MapFor';
+import Divider from 'material-ui/Divider';
+import {
+  enumerateLayouts,
+  enumerateExternalEvents,
+} from '../../ProjectManager/EnumerateProjectItems';
 import { type ParameterFieldProps } from './ParameterFieldProps.flow';
-import { defaultAutocompleteProps } from '../../../UI/AutocompleteProps';
+import { defaultAutocompleteProps } from '../../UI/AutocompleteProps';
 
 type State = {|
   focused: boolean,
   text: ?string,
 |};
 
-export default class LayerField extends Component<ParameterFieldProps, State> {
+export default class ExternalEventsField extends Component<
+  ParameterFieldProps,
+  State
+> {
   state = { focused: false, text: null };
 
-  _description: ?string;
+  _description: ?string = undefined;
+  _fullList: Array<{ text: string, value: string }> = [];
   _field: ?any;
-  _layersNames: Array<string> = [];
 
   constructor(props: ParameterFieldProps) {
     super(props);
@@ -33,22 +40,32 @@ export default class LayerField extends Component<ParameterFieldProps, State> {
   }
 
   componentWillReceiveProps(newProps: ParameterFieldProps) {
-    if (newProps.layout !== this.props.layout) {
+    if (newProps.project !== this.props.project) {
       this._loadNamesFrom(newProps);
     }
   }
 
   _loadNamesFrom(props: ParameterFieldProps) {
-    const layout = props.layout;
-    if (!layout) {
-      this._layersNames = [];
+    const { project } = props;
+    if (!project) {
       return;
     }
 
-    this._layersNames = mapFor(0, layout.getLayersCount(), i => {
-      const layer = layout.getLayerAt(i);
-      return layer.getName();
-    });
+    const externalEvents = enumerateExternalEvents(
+      project
+    ).map(externalEvents => ({
+      text: externalEvents.getName(),
+      value: externalEvents.getName(),
+    }));
+    const layouts = enumerateLayouts(project).map(layout => ({
+      text: layout.getName(),
+      value: layout.getName(),
+    }));
+    this._fullList = [
+      ...externalEvents,
+      { text: '', value: <Divider /> },
+      ...layouts,
+    ];
   }
 
   render() {
@@ -56,6 +73,7 @@ export default class LayerField extends Component<ParameterFieldProps, State> {
       <AutoComplete
         {...defaultAutocompleteProps}
         floatingLabelText={this._description}
+        id="external-events-field"
         searchText={this.state.focused ? this.state.text : this.props.value}
         onFocus={() => {
           this.setState({
@@ -85,10 +103,7 @@ export default class LayerField extends Component<ParameterFieldProps, State> {
           }
           this.focus(); // Keep the focus after choosing an item
         }}
-        dataSource={this._layersNames.map(layerName => ({
-          text: layerName || '(Base layer)',
-          value: `"${layerName}"`,
-        }))}
+        dataSource={this._fullList}
         openOnFocus={!this.props.isInline}
         ref={field => (this._field = field)}
       />
