+++ newIDE/app/src/EventsSheet/ParameterFields/OperatorField.js
