+++ newIDE/app/src/EventsSheet/ParameterFields/ParameterFieldProps.flow.js
@@ -2,8 +2,8 @@
 import {
   type ResourceSource,
   type ChooseResourceFunction,
-} from '../../../ResourcesList/ResourceSource.flow';
-import { type ResourceExternalEditor } from '../../../ResourcesList/ResourceExternalEditor.flow';
+} from '../../ResourcesList/ResourceSource.flow';
+import { type ResourceExternalEditor } from '../../ResourcesList/ResourceExternalEditor.flow';
 
 export type ParameterFieldProps = {|
   parameterMetadata?: gdParameterMetadata,
