+++ newIDE/app/src/EventsSheet/ParameterFields/SceneVariableField.js
@@ -1,14 +1,14 @@
 // @flow
 import * as React from 'react';
 import VariableField from './VariableField';
-import VariablesEditorDialog from '../../../VariablesList/VariablesEditorDialog';
+import VariablesEditorDialog from '../../VariablesList/VariablesEditorDialog';
 import { type ParameterFieldProps } from './ParameterFieldProps.flow';
 
 type State = {|
   editorOpen: boolean,
 |};
 
-export default class GlobalVariableField extends React.Component<
+export default class SceneVariableField extends React.Component<
   ParameterFieldProps,
   State
 > {
@@ -22,12 +22,12 @@ export default class GlobalVariableField extends React.Component<
   }
 
   render() {
-    const { project } = this.props;
+    const { layout } = this.props;
 
     return (
-      <div>
+      <React.Fragment>
         <VariableField
-          variablesContainer={project ? project.getVariables() : null}
+          variablesContainer={layout ? layout.getVariables() : null}
           parameterMetadata={this.props.parameterMetadata}
           value={this.props.value}
           onChange={this.props.onChange}
@@ -38,19 +38,20 @@ export default class GlobalVariableField extends React.Component<
           objectsContainer={this.props.objectsContainer}
         />
         {this.state.editorOpen &&
-          project && (
+          layout && (
             <VariablesEditorDialog
               open={this.state.editorOpen}
-              variablesContainer={project.getVariables()}
+              variablesContainer={layout.getVariables()}
               onCancel={() => this.setState({ editorOpen: false })}
               onApply={() => {
                 this.setState({ editorOpen: false });
                 if (this._field) this._field.forceUpdateVariables();
               }}
-              emptyExplanationMessage="Global variables are variables that are persisted across the scenes during the game."
+              emptyExplanationMessage="Scene variables can be used to store any value or text during the game."
+              emptyExplanationSecondMessage="For example, you can have a variable called Score representing the current score of the player."
             />
           )}
-      </div>
+      </React.Fragment>
     );
   }
 }
