+++ newIDE/app/src/stories/index.js
@@ -30,15 +30,18 @@ import CollisionMasksEditor from '../ObjectEditor/Editors/SpriteEditor/Collision
 import EmptyEditor from '../ObjectEditor/Editors/EmptyEditor';
 import ImageThumbnail from '../ResourcesList/ResourceThumbnail/ImageThumbnail';
 import ShapePainterEditor from '../ObjectEditor/Editors/ShapePainterEditor';
-import ExternalEventsField from '../EventsSheet/InstructionEditor/ParameterFields/ExternalEventsField';
-import LayerField from '../EventsSheet/InstructionEditor/ParameterFields/LayerField';
-import MouseField from '../EventsSheet/InstructionEditor/ParameterFields/MouseField';
-import SceneVariableField from '../EventsSheet/InstructionEditor/ParameterFields/SceneVariableField';
-import ObjectVariableField from '../EventsSheet/InstructionEditor/ParameterFields/ObjectVariableField';
-import KeyField from '../EventsSheet/InstructionEditor/ParameterFields/KeyField';
-import ExpressionField from '../EventsSheet/InstructionEditor/ParameterFields/ExpressionField';
-import StringField from '../EventsSheet/InstructionEditor/ParameterFields/StringField';
-import ColorExpressionField from '../EventsSheet/InstructionEditor/ParameterFields/ColorExpressionField';
+import ExternalEventsField from '../EventsSheet/ParameterFields/ExternalEventsField';
+import LayerField from '../EventsSheet/ParameterFields/LayerField';
+import MouseField from '../EventsSheet/ParameterFields/MouseField';
+import SceneVariableField from '../EventsSheet/ParameterFields/SceneVariableField';
+import ObjectVariableField from '../EventsSheet/ParameterFields/ObjectVariableField';
+import KeyField from '../EventsSheet/ParameterFields/KeyField';
+import ExpressionField from '../EventsSheet/ParameterFields/ExpressionField';
+import StringField from '../EventsSheet/ParameterFields/StringField';
+import ColorExpressionField from '../EventsSheet/ParameterFields/ColorExpressionField';
+import TrueFalseField from '../EventsSheet/ParameterFields/TrueFalseField';
+import YesNoField from '../EventsSheet/ParameterFields/YesNoField';
+import ForceMultiplierField from '../EventsSheet/ParameterFields/ForceMultiplierField';
 import ObjectsList from '../ObjectsList';
 import ObjectSelector from '../ObjectsList/ObjectSelector';
 import InstancePropertiesEditor from '../InstancesEditor/InstancePropertiesEditor';
@@ -59,7 +62,7 @@ import ResourcesLoader from '../ResourcesLoader';
 import VariablesList from '../VariablesList';
 import ExpressionSelector from '../EventsSheet/InstructionEditor/InstructionOrExpressionSelector/ExpressionSelector';
 import InstructionSelector from '../EventsSheet/InstructionEditor/InstructionOrExpressionSelector/InstructionSelector';
-import ParameterRenderingService from '../EventsSheet/InstructionEditor/ParameterRenderingService';
+import ParameterRenderingService from '../EventsSheet/ParameterRenderingService';
 import { ErrorFallbackComponent } from '../UI/ErrorBoundary';
 import { makeTestProject } from '../fixtures/TestProject';
 import CreateProfile from '../Profile/CreateProfile';
@@ -102,7 +105,7 @@ import ColorField from '../UI/ColorField';
 import EmptyMessage from '../UI/EmptyMessage';
 import BackgroundText from '../UI/BackgroundText';
 import i18n from '../UI/i18n';
-import ObjectField from '../EventsSheet/InstructionEditor/ParameterFields/ObjectField';
+import ObjectField from '../EventsSheet/ParameterFields/ObjectField';
 import { getInitialSelection } from '../EventsSheet/SelectionHandler';
 import EventsFunctionConfigurationEditor from '../EventsFunctionsExtensionEditor/EventsFunctionConfigurationEditor';
 import EventsFunctionsList from '../EventsFunctionsList';
@@ -513,6 +516,60 @@ storiesOf('ParameterFields', module)
           globalObjectsContainer={project}
           objectsContainer={testLayout}
           isInline
+        />
+      )}
+    />
+  ))
+  .add('TrueFalseField', () => (
+    <ValueStateHolder
+      initialValue={''}
+      render={(value, onChange) => (
+        <TrueFalseField
+          value={value}
+          onChange={onChange}
+          globalObjectsContainer={project}
+          objectsContainer={testLayout}
+          isInline
+        />
+      )}
+    />
+  ))
+  .add('YesNoField', () => (
+    <ValueStateHolder
+      initialValue={''}
+      render={(value, onChange) => (
+        <YesNoField
+          value={value}
+          onChange={onChange}
+          globalObjectsContainer={project}
+          objectsContainer={testLayout}
+          isInline
+        />
+      )}
+    />
+  ))
+  .add('ForceMultiplierField', () => (
+    <ValueStateHolder
+      initialValue={''}
+      render={(value, onChange) => (
+        <ForceMultiplierField
+          value={value}
+          onChange={onChange}
+          globalObjectsContainer={project}
+          objectsContainer={testLayout}
+        />
+      )}
+    />
+  ))
+  .add('ForceMultiplierField (with a deprecated value)', () => (
+    <ValueStateHolder
+      initialValue={'0.8'}
+      render={(value, onChange) => (
+        <ForceMultiplierField
+          value={value}
+          onChange={onChange}
+          globalObjectsContainer={project}
+          objectsContainer={testLayout}
         />
       )}
     />
